fix: agent runs
