+++ literalai/instrumentation/llamaindex/event_handler.py
@@ -1,20 +1,26 @@
-import logging
 import uuid
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union, cast
-from typing_extensions import TypedDict
-from llama_index.core.base.llms.types import MessageRole
-from llama_index.core.base.response.schema import Response, StreamingResponse
-from llama_index.core.instrumentation import get_dispatcher
+import logging
+from typing import TYPE_CHECKING, Dict, List, Optional, Union, cast
+
 from llama_index.core.instrumentation.event_handlers import BaseEventHandler
 from llama_index.core.instrumentation.events import BaseEvent
-from llama_index.core.base.llms.types import ChatMessage
-from llama_index.core.instrumentation.events.embedding import (
-    EmbeddingEndEvent,
-    EmbeddingStartEvent,
-)
+from pydantic import PrivateAttr
+
+from literalai.instrumentation.llamaindex.span_handler import LiteralSpanHandler
+from literalai.context import active_thread_var
 from llama_index.core.instrumentation.events.llm import (
     LLMChatEndEvent,
-    LLMChatStartEvent,
+)
+
+from llama_index.core.schema import QueryBundle
+from llama_index.core.instrumentation.events.agent import (
+    AgentChatWithStepStartEvent,
+    AgentChatWithStepEndEvent,
+    AgentRunStepStartEvent,
+    AgentRunStepEndEvent,
+)
+from llama_index.core.instrumentation.events.embedding import (
+    EmbeddingEndEvent,
 )
 
 from llama_index.core.instrumentation.events.query import QueryEndEvent, QueryStartEvent
@@ -22,18 +28,23 @@ from llama_index.core.instrumentation.events.retrieval import (
     RetrievalEndEvent,
     RetrievalStartEvent,
 )
+
+from llama_index.core.base.llms.types import MessageRole, ChatMessage
+from llama_index.core.base.response.schema import Response, StreamingResponse
+
+from llama_index.core.instrumentation.events.llm import (
+    LLMChatEndEvent,
+    LLMChatStartEvent,
+)
+
 from llama_index.core.instrumentation.events.synthesis import (
     SynthesizeEndEvent,
 )
-from llama_index.core.instrumentation.span import SimpleSpan
-from llama_index.core.instrumentation.span_handlers.base import BaseSpanHandler
-from llama_index.core.query_engine import RetrieverQueryEngine
+
 from llama_index.core.schema import NodeWithScore, QueryBundle, TextNode
 from openai.types import CompletionUsage
 from openai.types.chat import ChatCompletion
-from pydantic import PrivateAttr
 
-from literalai.context import active_thread_var
 from literalai.observability.generation import (
     ChatGeneration,
     GenerationMessage,
@@ -43,8 +54,6 @@ from literalai.observability.step import Step, StepType
 
 if TYPE_CHECKING:
     from literalai.client import LiteralClient
-
-literalai_uuid_namespace = uuid.UUID("05f6b2b5-a912-47bd-958f-98a9c4496322")
 
 
 def convert_message_role(role: MessageRole) -> GenerationMessageRole:
@@ -130,6 +139,14 @@ def create_generation(event: LLMChatStartEvent):
     )
 
 
+def print_blue(text: str):
+    print(f"\033[34m{text}\033[0m")
+
+
+def get_query(a: Union[str, QueryBundle]):
+    return a.query_str if isinstance(a, QueryBundle) else a
+
+
 class LiteralEventHandler(BaseEventHandler):
     """This class handles events coming from LlamaIndex."""
 
@@ -138,6 +155,7 @@ class LiteralEventHandler(BaseEventHandler):
     runs: Dict[str, List[Step]] = {}
     streaming_run_ids: List[str] = []
     _standalone_step_id: Optional[str] = None
+    open_runs: List[Step] = []
 
     class Config:
         arbitrary_types_allowed = True
@@ -176,11 +194,39 @@ class LiteralEventHandler(BaseEventHandler):
 
     def handle(self, event: BaseEvent, **kwargs) -> None:
         """Logic for handling event."""
+        tabs = self._span_handler.symbol * self._span_handler.tab_indent
+        print_blue(f"{tabs}{event.class_name()}")
         try:
             thread_id = self._span_handler.get_thread_id(event.span_id)
             run_id = self._span_handler.get_run_id(event.span_id)
 
             """The events are presented here roughly in chronological order"""
+            if isinstance(event, AgentChatWithStepStartEvent) or isinstance(
+                event, AgentRunStepStartEvent
+            ):
+                parent_run_id = None
+                if len(self.open_runs) > 0:
+                    parent_run_id = self.open_runs[-1].id
+
+                agent_run_id = str(uuid.uuid4())
+
+                run = self._client.start_step(
+                    name="Agent",
+                    type="run",
+                    id=agent_run_id,
+                    thread_id=thread_id,
+                    parent_id=parent_run_id,
+                )
+
+                self.open_runs.append(run)
+
+            if isinstance(event, AgentChatWithStepEndEvent) or isinstance(
+                event, AgentRunStepEndEvent
+            ):
+                step = self.open_runs.pop()
+                if step:
+                    step.end()
+
             if isinstance(event, QueryStartEvent):
                 active_thread = active_thread_var.get()
                 query = extract_query_from_bundle(event.query)
@@ -365,167 +411,3 @@ class LiteralEventHandler(BaseEventHandler):
                 return step
 
         return None
-
-
-class SpanEntry(TypedDict):
-    id: str
-    parent_id: Optional[str]
-    root_id: Optional[str]
-    is_run_root: bool
-
-
-class LiteralSpanHandler(BaseSpanHandler[SimpleSpan]):
-    """This class handles spans coming from LlamaIndex."""
-
-    spans: Dict[str, SpanEntry] = {}
-
-    @classmethod
-    def class_name(cls) -> str:
-        """Class name."""
-        return "LiteralSpanHandler"
-
-    def new_span(
-        self,
-        id_: str,
-        bound_args: Any,
-        instance: Optional[Any] = None,
-        parent_span_id: Optional[str] = None,
-        tags: Optional[Dict[str, Any]] = None,
-        **kwargs: Any,
-    ):
-        self.spans[id_] = {
-            "id": id_,
-            "parent_id": parent_span_id,
-            "root_id": None,
-            "is_run_root": self.is_run_root(instance, parent_span_id),
-        }
-
-        if parent_span_id is not None:
-            self.spans[id_]["root_id"] = self.get_root_span_id(parent_span_id)
-        else:
-            self.spans[id_]["root_id"] = id_
-
-    def is_run_root(
-        self, instance: Optional[Any], parent_span_id: Optional[str]
-    ) -> bool:
-        """Returns True if the span is of type RetrieverQueryEngine, and it has no run root in its parent chain"""
-        if not isinstance(instance, RetrieverQueryEngine):
-            return False
-
-        # Span is of correct type, we check that it doesn't have a run root in its parent chain
-        while parent_span_id:
-            parent_span = self.spans.get(parent_span_id)
-
-            if not parent_span:
-                parent_span_id = None
-                continue
-
-            if parent_span["is_run_root"]:
-                return False
-
-            parent_span_id = parent_span["parent_id"]
-
-        return True
-
-    def get_root_span_id(self, span_id: Optional[str]):
-        """Finds the root span and returns its ID"""
-        if not span_id:
-            return None
-
-        current_span = self.spans.get(span_id)
-
-        if current_span is None:
-            return None
-
-        while current_span["parent_id"] is not None:
-            current_span = self.spans.get(current_span["parent_id"])
-            if current_span is None:
-                return None
-
-        return current_span["id"]
-
-    def get_run_id(self, span_id: Optional[str]):
-        """Go up the span chain to find a run_root, return its ID (or None)"""
-        if not span_id:
-            return None
-
-        current_span = self.spans.get(span_id)
-
-        if current_span is None:
-            return None
-
-        while current_span:
-            if current_span["is_run_root"]:
-                return str(uuid.uuid5(literalai_uuid_namespace, current_span["id"]))
-
-            parent_id = current_span["parent_id"]
-
-            if parent_id:
-                current_span = self.spans.get(parent_id)
-            else:
-                current_span = None
-
-        return None
-
-    def get_thread_id(self, span_id: Optional[str]):
-        """Returns the root span ID as a thread ID"""
-        active_thread = active_thread_var.get()
-
-        if active_thread:
-            return active_thread.id
-
-        if span_id is None:
-            return None
-
-        current_span = self.spans.get(span_id)
-
-        if current_span is None:
-            return None
-
-        root_id = current_span["root_id"]
-
-        if not root_id:
-            return None
-
-        root_span = self.spans.get(root_id)
-
-        if root_span is None:
-            # span is already the root, uuid its own id
-            return str(uuid.uuid5(literalai_uuid_namespace, span_id))
-        else:
-            # uuid the id of the root span
-            return str(uuid.uuid5(literalai_uuid_namespace, root_span["id"]))
-
-    def prepare_to_exit_span(
-        self,
-        id_: str,
-        bound_args: Any,
-        instance: Optional[Any] = None,
-        result: Optional[Any] = None,
-        **kwargs: Any,
-    ):
-        """Logic for preparing to exit a span."""
-        if id_ in self.spans:
-            del self.spans[id_]
-
-    def prepare_to_drop_span(
-        self,
-        id_: str,
-        bound_args: Any,
-        instance: Optional[Any] = None,
-        err: Optional[BaseException] = None,
-        **kwargs: Any,
-    ):
-        """Logic for preparing to drop a span."""
-        if id_ in self.spans:
-            del self.spans[id_]
-
-
-def instrument_llamaindex(client: "LiteralClient"):
-    root_dispatcher = get_dispatcher()
-    span_handler = LiteralSpanHandler()
-    event_handler = LiteralEventHandler(
-        literal_client=client, llama_index_span_handler=span_handler
-    )
-    root_dispatcher.add_event_handler(event_handler)
-    root_dispatcher.add_span_handler(span_handler)
